+++ src/main/java/codesquad/webserver/handler/RouterHandler.java
@@ -1,0 +1,9 @@
+package codesquad.webserver.handler;
+
+import codesquad.http.HttpRequest;
+import codesquad.http.HttpResponse;
+
+public interface RouterHandler {
+    boolean support(HttpRequest httpRequest);
+    HttpResponse handle(HttpRequest httpRequest);
+}
